go/consensus/tendermint/apps/registry: Move transactions to its own file
